Adding support for a lot of operations

+++ include/data_asm.h
@@ -11,6 +11,6 @@ void assemble_dd(size_t value, ByteStream *stream); // 32 bits
 void assemble_dw(size_t value, ByteStream *stream); // 16 bits
 void assemble_db(size_t value, ByteStream *stream); // 8 bits
 
-void handle_data_directive(ByteStream *stream, directive dir, const char *data, int *pos);
+void handle_data_directive(ByteStream *stream, directive dir, const char *data, size_t *pos);
 
 #endif
