+++ include/utils.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-size_t getNum(const char *string, int *pos);
-size_t atob(const char *string, int *pos); // la usare para convertir binario
+size_t getNum(const char *string, size_t *pos);
+size_t atob(const char *string, size_t *pos); // la usare para convertir binario
 
 #endif
